requirements.txt
numpy
pandas<2.0.0
Flask
gunicorn
flask_bcrypt
flask_sqlalchemy
flask_login
flask_cors
flask_wtf
python-dotenv
pinecone-client